update logout logic (#23)
* feat: fix login

* feat: implement logout loading state

+++ src/layouts/Menu.tsx
@@ -2,12 +2,12 @@ import logo from "../assets/logo/logo.svg";
 import bookmarksIcon from "../assets/icons/bookmarks.svg";
 import categoriesIcon from "../assets/icons/categories.svg";
 import logoutIcon from "../assets/icons/logout.svg";
+import { MenuButton } from "../components/buttons";
+import { LoadingModal } from "../components/modals";
 
 import { ActiveContext } from "../pages/User";
-import { MenuButton } from "../components/buttons";
 import { logoutUser } from "../api";
 import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
 
 type MenuProps = {
   activeTab: ActiveContext;
@@ -23,8 +23,7 @@ export type MenuButtonProps = {
 
 export function Menu({ activeTab }: MenuProps) {
   const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-  const navigate = useNavigate();
-
+  const [loggingOut, setLoggingOut] = useState(false);
   useEffect(() => {
     function checkWindowWidth() {
       setWindowWidth(window.innerWidth);
@@ -46,13 +45,20 @@ export function Menu({ activeTab }: MenuProps) {
     activeTab.setBookmarksActive(false);
   }
   function logout() {
+    let successful = null;
     const controller = new AbortController();
     const logout = async () => {
-      await logoutUser(controller);
+      setLoggingOut(true);
+      successful = await logoutUser(controller);
+      if (successful) {
+        setLoggingOut(false);
+        localStorage.clear();
+        location.href = "http://192.168.1.9:5173/login";
+      } else {
+        alert("There was a problem logging you out. Please try again.");
+      }
     };
     logout();
-    localStorage.clear();
-    location.reload();
   }
   const menuButtons: MenuButtonProps[] = [
     {
@@ -60,20 +66,17 @@ export function Menu({ activeTab }: MenuProps) {
       text: "Home",
       active: activeTab.bookmarksActive,
       toggle: toggleBookmarks,
-      // showText: activeTab.inView,
     },
     {
       icon: categoriesIcon,
       text: "Categories",
       active: activeTab.categoriesActive,
       toggle: toggleCategories,
-      // showText: activeTab.inView,
     },
     {
       icon: logoutIcon,
       text: "Logout",
       toggle: logout,
-      // showText: activeTab.inView,
     },
   ];
 
@@ -83,39 +86,45 @@ export function Menu({ activeTab }: MenuProps) {
   }
 
   return (
-    <div
-      className={`menu__container ${addMobileClass(
-        `menu__mobile__container`
-      )} `}
-    >
-      <div className={`menu ${addMobileClass("menu__mobile")}`}>
-        <div className="menu__logo__container">
-          <div className="menu__logo">
-            <img src={logo} alt="logo" />
-            <h1>Sorta</h1>
+    <>
+      <div
+        className={`menu__container ${addMobileClass(
+          `menu__mobile__container`
+        )} `}
+      >
+        <div className={`menu ${addMobileClass("menu__mobile")}`}>
+          <div className="menu__logo__container">
+            <div className="menu__logo">
+              <img src={logo} alt="logo" />
+              <h1>Sorta</h1>
+            </div>
+          </div>
+          <div
+            className={`menu__buttons ${addMobileClass(
+              "menu__mobile__buttons"
+            )}`}
+          >
+            {menuButtons.map(({ icon, text, active, toggle }, index) => (
+              <MenuButton
+                icon={icon}
+                text={text}
+                active={active}
+                toggle={toggle}
+                // showText={showText}s
+                key={index}
+                mobileClass={addMobileClass(
+                  `menu__mobile__button ${
+                    active ? "menu__mobile__button--active" : ""
+                  } `
+                )}
+              />
+            ))}
           </div>
         </div>
-        <div
-          className={`menu__buttons ${addMobileClass("menu__mobile__buttons")}`}
-        >
-          {menuButtons.map(({ icon, text, active, toggle }, index) => (
-            <MenuButton
-              icon={icon}
-              text={text}
-              active={active}
-              toggle={toggle}
-              // showText={showText}s
-              key={index}
-              mobileClass={addMobileClass(
-                `menu__mobile__button ${
-                  active ? "menu__mobile__button--active" : ""
-                } `
-              )}
-            />
-          ))}
-        </div>
       </div>
-    </div>
+
+      {loggingOut && <LoadingModal />}
+    </>
   );
 }
